+++ src/constants/actions.js
@@ -3,10 +3,10 @@ export const LOCAL_SET_APP = 'LOCAL_SET_APP';
 export const LOCAL_REMOVE_APP = 'LOCAL_REMOVE_APP';
 
 // Apps
-export const TOP_CHARTS_RESET = 'TOP_CHARTS_RESET';
-export const TOP_CHARTS_GET_FAILED = 'TOP_CHARTS_GET_FAILED';
-export const TOP_CHARTS_GET_REQUEST = 'TOP_CHARTS_GET_REQUEST';
-export const TOP_CHARTS_GET_SUCCESS = 'TOP_CHARTS_GET_SUCCESS';
+export const POPULAR_APPS_RESET = 'POPULAR_APPS_RESET';
+export const POPULAR_APPS_GET_FAILED = 'POPULAR_APPS_GET_FAILED';
+export const POPULAR_APPS_GET_REQUEST = 'POPULAR_APPS_GET_REQUEST';
+export const POPULAR_APPS_GET_SUCCESS = 'POPULAR_APPS_GET_SUCCESS';
 
 // User
 export const USER_GET_REQUEST = 'USER_GET_REQUEST';
