+++ src/constants/routes.js
@@ -1,3 +1,3 @@
 export const ROUTE_INSTALLED_APPS = 'ROUTE_INSTALLED_APPS';
-export const ROUTE_TOP_CHARTS = 'ROUTE_TOP_CHARTS';
+export const ROUTE_POPULAR_APPS = 'ROUTE_POPULAR_APPS';
 export const ROUTE_SEARCH = 'ROUTE_SEARCH';
