+++ src/constants/strings.js
@@ -1,11 +1,4 @@
 export const STRING_ABOUT = 'About';
-export const STRING_ACCOUNT = 'Account';
-export const STRING_APP_NAME = 'Name';
-export const STRING_APP_NAME_PLACEHOLDER = 'e.g. Gmail';
-export const STRING_APP_URL = 'URL';
-export const STRING_APP_URL_PLACEHOLDER = 'e.g. gmail.com';
-export const STRING_AUTH_REQUIRED = 'Authentication Is Required';
-export const STRING_AUTH_REQUIRED_DESC = 'You need to sign in to access this page.';
 export const STRING_BACK = 'Back';
 export const STRING_BROWSING_DATA_REMOVED = 'All of your browsing data for {appName} will be removed and can\'t be recovered.';
 export const STRING_CANCEL = 'Cancel';
@@ -13,65 +6,22 @@ export const STRING_CHECK_FOR_UPDATES = 'Check for updates';
 export const STRING_CHECKING_FOR_UPDATES = 'Checking for updates...';
 export const STRING_CLEAR = 'Clear';
 export const STRING_CLOSE = 'Close';
-export const STRING_CONFIRM_NEW_PASSWORD = 'Confirm New Password';
-export const STRING_CONFIRM_NEW_PASSWORD_PLACEHOLDER = 'Re-enter your new password';
-export const STRING_CONTINUE_WITHOUT_AN_ACCOUNT = 'Continue without an account';
-export const STRING_CREATE_ACCOUNT = 'Create an account';
-export const STRING_CURRENT_PASSWORD = 'Current Password';
-export const STRING_CURRENT_PASSWORD_PLACEHOLDER = 'Enter your current password';
-export const STRING_DISPLAY_NAME = 'Display Name';
-export const STRING_DISPLAY_NAME_PLACEHOLDER = 'Enter a name to go by';
-export const STRING_EMAIL = 'Email';
-export const STRING_EMAIL_PLACEHOLDER = 'Enter your email';
 export const STRING_FAILED_TO_CONNECT = 'Failed to Connect to Server';
 export const STRING_FAILED_TO_CONNECT_DESC = 'Please check your Internet connection.';
 export const STRING_FAILED_TO_INSTALL = 'We\'re sorry. WebCatalog has failed to install {name}.';
 export const STRING_FAILED_TO_UPDATE = 'We\'re sorry. WebCatalog has failed to update {name}.';
-export const STRING_FILTER = 'Filter';
-export const STRING_FORGOT_PASSWORD = 'Forgot your password?';
-export const STRING_HELP_AND_FEEDBACK = 'Help & Feedback';
 export const STRING_INSTALL = 'Install';
 export const STRING_INSTALLED_APPS = 'Installed Apps';
 export const STRING_INSTALLING = 'Installing...';
-export const STRING_LOG_IN = 'Sign in';
-export const STRING_LOG_IN_WITH_GOOGLE = 'Sign in with Google';
-export const STRING_LOG_OUT = 'Log out';
-export const STRING_MENU = 'Menu';
-export const STRING_NEW_APPS = 'New Apps';
-export const STRING_NEW_APPS_IN_CATEGORY = 'New Apps in {category}';
-export const STRING_NEW_PASSWORD = 'New Password';
-export const STRING_NEW_PASSWORD_PLACEHOLDER = 'Enter your new password';
-export const STRING_NO_APPS = 'No Apps';
 export const STRING_NO_APPS_DESC = 'Your previously installed apps will show up here.';
 export const STRING_NO_INSTALLED_APPS = 'No Installed Apps';
 export const STRING_NO_INSTALLED_APPS_DESC = 'Your installed apps on this machine will show up here.';
-export const STRING_NO_RESULTS = 'No Results';
-export const STRING_NO_RESULTS_HINT = 'Try a synonym, a domain or more general keyword';
 export const STRING_OPEN = 'Open';
 export const STRING_OPEN_ABOUT_DIALOG = 'Open About dialog';
-export const STRING_PASSWORD = 'Password';
 export const STRING_PRIVACY_POLICY = 'Privacy Policy';
-export const STRING_PROFILE = 'Profile';
-export const STRING_REFRESH = 'Refresh';
 export const STRING_RELEASE_NOTES = 'Release Notes';
 export const STRING_UPDATE_AND_RELAUNCH = 'Update & relaunch';
-export const STRING_SAVE = 'Save';
-export const STRING_SAVING = 'Saving';
-export const STRING_SEARCH = 'Search';
-export const STRING_SEARCH_APPS = 'Search Apps';
-export const STRING_SEND = 'Send';
-export const STRING_SENDING = 'Sending';
-export const STRING_SUBMIT = 'Submit';
-export const STRING_SUBMIT_APP = 'Submit App';
-export const STRING_SUBMITTING = 'Submitting...';
 export const STRING_TERMS = 'Terms of Service';
-export const STRING_TOP_APPS = 'Top Apps';
-export const STRING_TOP_APPS_IN_CATEGORY = 'Top Apps in {category}';
-export const STRING_TOP_CHARTS = 'Top Charts';
-export const STRING_TRY_AGAIN = 'Try again';
-export const STRING_UNINSTALL = 'Uninstall';
-export const STRING_UNINSTALLING = 'Uninstalling...';
-export const STRING_UNINSTALLING_APP_NAME = 'Uninstall {appName}?';
 export const STRING_UPDATE = 'Update';
 export const STRING_UPDATE_AVAILABLE = 'An update is available. It\'ll be downloaded soon.';
 export const STRING_UPDATE_AVAILABLE_LINUX = 'An update is available. Please go to our website to download.';
@@ -86,7 +36,9 @@ export const STRING_WHATS_NEW = 'what\'s new';
 export const STRING_UPDATE_ALL = 'Update all';
 export const STRING_UPDATING = 'Updating...';
 export const STRING_UPDATES_AVAILABLE = 'Updates available';
-export const STRING_CANT_FIND_YOUR_FAVORITE_APP = 'Can\'t find your favorite apps?';
-export const STRING_LET_US_KNOW = 'Let us know';
 export const STRING_GO_TO_THE_WEBSITE = 'Go to the website';
 export const STRING_VIEW = 'View';
+export const STRING_POPULAR_APPS = 'Popular Apps';
+export const STRING_UNINSTALL = 'Uninstall';
+export const STRING_UNINSTALLING = 'Uninstalling...';
+export const STRING_UNINSTALLING_APP_NAME = 'Uninstall {appName}?';
