+++ src/reducers/pages/popular-apps/reducers.js
